Move supported version check from driver to version pkg
