+++ operators/pkg/controller/elasticsearch/version/supported_versions.go
@@ -20,6 +20,27 @@ type LowestHighestSupportedVersions struct {
 	HighestSupportedVersion version.Version
 }
 
+func SupportedVersions(v version.Version) *LowestHighestSupportedVersions {
+	switch v.Major {
+	case 6:
+		return &LowestHighestSupportedVersions{
+			// Min. version is 6.7.0 for now. Will be 6.8.0 soon.
+			LowestSupportedVersion: version.MustParse("6.7.0"),
+			// higher may be possible, but not proven yet, lower may also be a requirement...
+			HighestSupportedVersion: version.MustParse("6.99.99"),
+		}
+	case 7:
+		return &LowestHighestSupportedVersions{
+			// 6.7.0 is the lowest wire compatibility version for 7.x
+			LowestSupportedVersion: version.MustParse("6.7.0"),
+			// higher may be possible, but not proven yet, lower may also be a requirement...
+			HighestSupportedVersion: version.MustParse("7.99.99"),
+		}
+	default:
+		return nil
+	}
+}
+
 // VerifySupportsExistingPods checks the given pods against the supported version range in lh.
 func (lh LowestHighestSupportedVersions) VerifySupportsExistingPods(
 	pods []corev1.Pod,
